Moved marathon serviceinit functions to instance.py
