+++ paasta_tools/api/views/instance.py
@@ -61,8 +61,7 @@ from paasta_tools.kubernetes_tools import KubeClient
 from paasta_tools.kubernetes_tools import KubernetesDeploymentConfig
 from paasta_tools.long_running_service_tools import LongRunningServiceConfig
 from paasta_tools.long_running_service_tools import ServiceNamespaceConfig
-from paasta_tools.marathon_serviceinit import get_marathon_dashboard_links
-from paasta_tools.marathon_serviceinit import get_short_task_id
+from paasta_tools.marathon_tools import get_short_task_id
 from paasta_tools.mesos.task import Task
 from paasta_tools.mesos_tools import (
     get_cached_list_of_not_running_tasks_from_frameworks,
@@ -1068,3 +1067,17 @@ def add_executor_info(task):
 def add_slave_info(task):
     task._Task__items["slave"] = a_sync.block(task.slave)._MesosSlave__items.copy()
     return task
+
+
+def get_marathon_dashboard_links(marathon_clients, system_paasta_config):
+    """Return a dict of marathon clients and their corresponding dashboard URLs"""
+    cluster = system_paasta_config.get_cluster()
+    try:
+        links = (
+            system_paasta_config.get_dashboard_links().get(cluster).get("Marathon RO")
+        )
+    except KeyError:
+        pass
+    if isinstance(links, list) and len(links) >= len(marathon_clients.current):
+        return {client: url for client, url in zip(marathon_clients.current, links)}
+    return None
