+++ paasta_tools/marathon_tools.py
@@ -1615,3 +1615,8 @@ def take_up_slack(client: MarathonClient, app: MarathonApp) -> None:
             % (app.id, app.instances, app.instances - slack)
         )
         client.scale_app(app_id=app.id, instances=(app.instances - slack), force=True)
+
+
+def get_short_task_id(task_id):
+    """Return just the Marathon-generated UUID of a Mesos task id."""
+    return task_id.split(MESOS_TASK_SPACER)[-1]
