+++ tests/api/test_instance.py
@@ -1012,3 +1012,41 @@ def test_kubernetes_instance_status_evicted_nodes(
         {}, "fake-svc", "fake-inst", 0, False, "kubernetes"
     )
     assert instance_status["evicted_count"] == 1
+
+
+def test_get_marathon_dashboard_links():
+    system_paasta_config = SystemPaastaConfig(
+        config={
+            "cluster": "fake_cluster",
+            "dashboard_links": {
+                "fake_cluster": {
+                    "Marathon RO": [
+                        "http://marathon",
+                        "http://marathon1",
+                        "http://marathon2",
+                    ]
+                }
+            },
+        },
+        directory="/fake/config",
+    )
+
+    marathon_clients = mock.Mock(current=["client", "client1", "client2"])
+    assert instance.get_marathon_dashboard_links(
+        marathon_clients, system_paasta_config
+    ) == {
+        "client": "http://marathon",
+        "client1": "http://marathon1",
+        "client2": "http://marathon2",
+    }
+
+    marathon_clients = mock.Mock(current=["client", "client1", "client2", "client3"])
+    assert (
+        instance.get_marathon_dashboard_links(marathon_clients, system_paasta_config)
+        is None
+    )
+
+    marathon_clients = mock.Mock(current=["client", "client1"])
+    assert instance.get_marathon_dashboard_links(
+        marathon_clients, system_paasta_config
+    ) == {"client": "http://marathon", "client1": "http://marathon1"}
